Now allows list and strings, documentation
Now allows for both, enable and disable function, to either give a
string or a list as the wanted links, for both the first and second
links. Also adds now first and second links to the ACM for disabling
collision detections, instead of only second links. Finally, added some
documentation.