Rough namespace fix
By providing "/" as the namepsace parameter for both the
MoveGroupCommander and the PLanningSceneInterface, it is now possible to
use them, when they lie in a namespace other than global while the
move_group and planning_scene_interface nodes still lie in the global
namespace. This is accomplished by checking for this specific namespace
parameter in the pytho nodes and then setting the name for the node
handle for the c++ node to a specific value. If this value is detected
in the c++ node, a "/" is added before the otherwise relative names for
all required services, topics, etc.